fix: find safe entity wormhole exit positions (#1601)
* fix: find safe entity wormhole exit positions

* chore: code cleanup

+++ src/main/java/com/klikli_dev/occultism/client/render/blockentity/GoldenSacrificialBowlRenderer.java
@@ -31,15 +31,12 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.math.Axis;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.LevelRenderer;
-import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider.Context;
-import net.minecraft.client.renderer.feature.ModelFeatureRenderer.CrumblingOverlay;
-import net.minecraft.client.renderer.item.ItemModelResolver;
 import net.minecraft.client.renderer.SubmitNodeCollector;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderer;
-import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider;
-import net.minecraft.client.renderer.blockentity.state.BlockEntityRenderState;
+import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider.Context;
 import net.minecraft.client.renderer.entity.EntityRenderDispatcher;
-import net.minecraft.client.renderer.feature.ModelFeatureRenderer;
+import net.minecraft.client.renderer.feature.ModelFeatureRenderer.CrumblingOverlay;
+import net.minecraft.client.renderer.item.ItemModelResolver;
 import net.minecraft.client.renderer.item.ItemStackRenderState;
 import net.minecraft.client.renderer.state.level.CameraRenderState;
 import net.minecraft.client.renderer.texture.OverlayTexture;
@@ -53,13 +50,13 @@ import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemDisplayContext;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.phys.Vec3;
+import org.jspecify.annotations.Nullable;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
-import net.minecraft.world.level.block.state.properties.BlockStateProperties;
-import net.minecraft.world.phys.Vec3;
-import org.jspecify.annotations.Nullable;
 
 public class GoldenSacrificialBowlRenderer implements BlockEntityRenderer<SacrificialBowlBlockEntity, GoldenSacrificialBowlRenderState> {
 
@@ -160,7 +157,7 @@ public class GoldenSacrificialBowlRenderer implements BlockEntityRenderer<Sacrif
                         BlockPos previewPos = blockEntity.getBlockPos().relative(facing, 2).above(3);
                         entity.setPos(previewPos.getX() + 0.5, previewPos.getY(), previewPos.getZ() + 0.5);
 
-                        float maxSize = (float) Math.max(entity.getBbWidth(), Math.max(entity.getBbHeight(), entity.getBbWidth()));
+                        float maxSize = Math.max(entity.getBbWidth(), Math.max(entity.getBbHeight(), entity.getBbWidth()));
                         renderState.sacrificeEntityScale = maxSize > 0 ? 0.5F / maxSize : 1.0F;
                         renderState.sacrificeEntityRenderState = Minecraft.getInstance().getEntityRenderDispatcher().extractEntity(entity, partialTick);
                         renderState.sacrificeEntityRenderState.lightCoords = LevelRenderer.getLightCoords(blockEntity.getLevel(), previewPos);
